Add VirtuosoConnector
#392
Add a new connector for Virtuoso which supports the methods of the ITransactionalStorage interface.
Note that the method called by SaveGraph always commits the new graph regardless of the current transaction state, so Commit/Rollback really only applies to UpdateGraph.
