+++ Testing/dotNetRDF.Connectors.Virtuoso.Tests/VirtuosoConnectorAsync.cs
@@ -1,0 +1,14 @@
+﻿using Xunit;
+
+namespace VDS.RDF.Storage
+{
+    [Collection("Virtuoso")]
+    public class VirtuosoConnectorAsync
+        : BaseAsyncTests
+    {
+        protected override IAsyncStorageProvider GetAsyncProvider()
+        {
+            return VirtuosoConnectorTest.GetConnection();
+        }
+    }
+}
