+++ Testing/dotNetRDF.Connectors.Virtuoso.Tests/VirtuosoConnectorPersistentTripleStoreTests.cs
@@ -1,0 +1,12 @@
+﻿using Xunit;
+
+namespace VDS.RDF.Storage;
+
+[Collection("Virtuoso")]
+public class VirtuosoConnectorPersistentTripleStoreTests : PersistentTripleStoreTests
+{
+    protected override IStorageProvider GetStorageProvider()
+    {
+        return VirtuosoConnectorTest.GetConnection();
+    }
+}
